Fix hook engine crash when target shares page with NtProtectVirtualMemory
Trampoline pages are now allocated as RWX (immediately callable) then
downgraded to RX via make_executable(). Target patching uses RWX instead
of RW because NtClose and NtProtectVirtualMemory can share the same
ntdll page — removing execute permission mid-syscall caused a segfault.

Also restores missing codegen methods (blr, ret, brk, nop) and fixes
minor IR/codegen warnings from prior session.

+++ lib/ir/ir_lift.cpp
@@ -68,7 +68,7 @@ static uint8_t arr_vec_size(Register r) {
 // Template interpreters
 // ============================================================================
 
-static uint16_t temp_idx;
+static uint32_t temp_idx;
 
 static VarNode next_temp(uint8_t sz) {
     return VarNode::temp(temp_idx++, sz);
